chore:#404 generateImage 외부 인프라/AI 패키지 분리
