--- src/main/java/or/sopt/houme/domain/gemini/service/GeminiImageService.java
@@ -1,7 +1,0 @@
-package or.sopt.houme.domain.gemini.service;
-
-import or.sopt.houme.global.dto.ImageUploadResponseDTO;
-
-public interface GeminiImageService {
-    ImageUploadResponseDTO createImage(String prompt);
-}
