+++ src/main/java/or/sopt/houme/domain/generateImage/infrastructure/gemini/service/GeminiImageService.java
@@ -1,0 +1,7 @@
+package or.sopt.houme.domain.generateImage.infrastructure.gemini.service;
+
+import or.sopt.houme.global.dto.ImageUploadResponseDTO;
+
+public interface GeminiImageService {
+    ImageUploadResponseDTO createImage(String prompt);
+}
