+++ src/main/java/or/sopt/houme/domain/generateImage/infrastructure/openai/service/FastApiService.java
@@ -1,0 +1,8 @@
+package or.sopt.houme.domain.generateImage.infrastructure.openai.service;
+
+import or.sopt.houme.domain.generateImage.service.prompt.dto.PromptRequestDTO;
+import or.sopt.houme.global.dto.ImageUploadResponseDTO;
+
+public interface FastApiService {
+    ImageUploadResponseDTO getImageByFastApi(PromptRequestDTO request);
+}
