+++ src/main/java/or/sopt/houme/domain/generateImage/infrastructure/openai/service/OpenAiService.java
@@ -1,0 +1,10 @@
+package or.sopt.houme.domain.generateImage.infrastructure.openai.service;
+
+import or.sopt.houme.domain.generateImage.presentation.openai.controller.dto.OpenAiRequest;
+import or.sopt.houme.global.dto.ImageUploadResponseDTO;
+
+public interface OpenAiService {
+    ImageUploadResponseDTO createImage(String prompt);
+
+    byte[] getGptImage(OpenAiRequest request);
+}
