+++ src/main/java/or/sopt/houme/domain/generateImage/service/prompt/PromptService.java
@@ -1,0 +1,8 @@
+package or.sopt.houme.domain.generateImage.service.prompt;
+
+import or.sopt.houme.domain.generateImage.service.prompt.dto.PromptRequestDTO;
+
+public interface PromptService {
+
+    String makePrompt(PromptRequestDTO promptRequestDTO);
+}
