+++ src/main/java/or/sopt/houme/domain/generateImage/service/prompt/dto/PromptFurnitureListDTO.java
@@ -1,0 +1,12 @@
+package or.sopt.houme.domain.generateImage.service.prompt.dto;
+
+import java.util.List;
+
+public record PromptFurnitureListDTO(
+        List<Long> furnitureTagIds
+) {
+
+    public static PromptFurnitureListDTO of(List<Long> furnitureIds) {
+        return new PromptFurnitureListDTO(furnitureIds);
+    }
+}
