--- src/main/java/or/sopt/houme/domain/openai/controller/dto/OpenAiResponse.java
@@ -1,26 +1,0 @@
-package or.sopt.houme.domain.openai.controller.dto;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.*;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Data
-@JsonIgnoreProperties(ignoreUnknown = true)
-@AllArgsConstructor
-@NoArgsConstructor
-public class OpenAiResponse {
-    private List<ImageData> data;
-
-    @Data
-    @JsonIgnoreProperties(ignoreUnknown = true)
-    @AllArgsConstructor
-    @NoArgsConstructor
-    public static class ImageData {
-        private String url;
-        private String revised_prompt;
-        private String b64_json;
-    }
-}
