--- src/main/java/or/sopt/houme/domain/openai/service/FastApiService.java
@@ -1,8 +1,0 @@
-package or.sopt.houme.domain.openai.service;
-
-import or.sopt.houme.domain.prompt.dto.PromptRequestDTO;
-import or.sopt.houme.global.dto.ImageUploadResponseDTO;
-
-public interface FastApiService {
-    ImageUploadResponseDTO getImageByFastApi(PromptRequestDTO request);
-}
