--- src/main/java/or/sopt/houme/domain/prompt/service/PromptService.java
@@ -1,8 +1,0 @@
-package or.sopt.houme.domain.prompt.service;
-
-import or.sopt.houme.domain.prompt.dto.PromptRequestDTO;
-
-public interface PromptService {
-
-    String makePrompt(PromptRequestDTO promptRequestDTO);
-}
